refactor: split async task helpers into domain-named task modules
Applies decision 0010: async_helpers.rs (1,115 lines, a mechanism-named
file spanning six domains) becomes app/tasks/ with one module per domain
and the shared run_off_ui_thread idiom in mod.rs. Pure moves; test
modules travel with their functions.
