--- crates/vibez-ui/src/app/async_helpers.rs
@@ -1,1115 +1,0 @@
-//! Async decode/save/load/dropbox/export task helpers.
-//! Split from app/mod.rs; re-exported into the app module.
-
-use super::*;
-
-pub(super) async fn decode_local_for_preview_async(
-    path: PathBuf,
-) -> Result<crate::message::AnalysedBrowserAudio, String> {
-    tokio::task::spawn_blocking(move || {
-        let name = path
-            .file_name()
-            .map(|value| value.to_string_lossy().into_owned())
-            .unwrap_or_default();
-        let audio = file_io::decode_audio_file(&path).map_err(|error| error.to_string())?;
-        let loop_fit = crate::warp::analyse_loop_tempo(&audio, &name);
-        Ok(crate::message::AnalysedBrowserAudio {
-            audio: Arc::new(audio),
-            loop_fit,
-        })
-    })
-    .await
-    .map_err(|error| format!("preview decode task failed: {error}"))?
-}
-
-pub(super) async fn decode_file_async(
-    path: PathBuf,
-) -> Result<vibez_core::audio_buffer::DecodedAudio, String> {
-    tokio::task::spawn_blocking(move || {
-        file_io::decode_audio_file(&path).map_err(|e| e.to_string())
-    })
-    .await
-    .map_err(|e| format!("decode task failed: {e}"))?
-}
-
-pub(super) async fn analyse_browser_audio_async(
-    audio: vibez_core::audio_buffer::DecodedAudio,
-    source_name: String,
-) -> Result<crate::message::AnalysedBrowserAudio, String> {
-    analyse_browser_audio_with_cached_metadata_async(audio, source_name, None).await
-}
-
-async fn analyse_browser_audio_with_cached_metadata_async(
-    audio: vibez_core::audio_buffer::DecodedAudio,
-    source_name: String,
-    cached_metadata: Option<vibez_dropbox::DerivedMetadata>,
-) -> Result<crate::message::AnalysedBrowserAudio, String> {
-    tokio::task::spawn_blocking(move || {
-        let loop_fit = match cached_metadata {
-            Some(metadata) => crate::warp::fit_loop_tempo(
-                audio.num_frames(),
-                audio.sample_rate as f64,
-                &source_name,
-                metadata.bpm,
-            ),
-            None => crate::warp::analyse_loop_tempo(&audio, &source_name),
-        };
-        crate::message::AnalysedBrowserAudio {
-            audio: Arc::new(audio),
-            loop_fit,
-        }
-    })
-    .await
-    .map_err(|error| format!("Browser analysis task failed: {error}"))
-}
-
-pub(super) async fn decode_and_analyse_async(
-    path: PathBuf,
-    source_name: String,
-) -> Result<crate::message::AnalysedBrowserAudio, String> {
-    let audio = decode_file_async(path).await?;
-    analyse_browser_audio_async(audio, source_name).await
-}
-
-pub(super) async fn decode_analyse_and_stage_local_async(
-    path: PathBuf,
-) -> Result<(crate::message::AnalysedBrowserAudio, MediaSourceRef), String> {
-    let name = path
-        .file_name()
-        .map(|value| value.to_string_lossy().into_owned())
-        .unwrap_or_default();
-    let (audio, source) = decode_and_stage_local_async(path).await?;
-    let analysed = analyse_browser_audio_async(audio, name).await?;
-    Ok((analysed, source))
-}
-
-pub(super) async fn decode_and_stage_local_async(
-    path: PathBuf,
-) -> Result<(vibez_core::audio_buffer::DecodedAudio, MediaSourceRef), String> {
-    tokio::task::spawn_blocking(move || {
-        // One read feeds both the decoder and the staging copy, so the
-        // engine can never play different bytes than the project commits
-        // (the file could be replaced between two independent reads).
-        let content = std::fs::read(&path).map_err(|error| error.to_string())?;
-        let extension = path
-            .extension()
-            .map(|value| value.to_string_lossy().into_owned());
-        let audio = file_io::decode_audio_cursor(
-            std::io::Cursor::new(content.clone()),
-            extension.as_deref(),
-        )
-        .map_err(|error| error.to_string())?;
-        let file_name = path
-            .file_name()
-            .map(|name| name.to_string_lossy().into_owned())
-            .unwrap_or_else(|| "project-media".to_string());
-        let source =
-            vibez_project::project_format_v1::stage_local_content(&path, &file_name, &content)
-                .map_err(|error| error.to_string())?;
-        Ok((audio, source))
-    })
-    .await
-    .map_err(|error| format!("decode/stage task failed: {error}"))?
-}
-
-pub(super) async fn save_project_async(
-    path: PathBuf,
-    source_path: Option<PathBuf>,
-    project: Project,
-) -> Result<ProjectSaveResult, String> {
-    tokio::task::spawn_blocking(move || {
-        let is_v1_destination = path
-            .extension()
-            .is_some_and(|extension| extension.eq_ignore_ascii_case("vzp"));
-        if is_v1_destination {
-            let v1_source = source_path.as_deref().filter(|source| {
-                vibez_project::project_format_v1::detect_project_format(source).is_ok_and(
-                    |format| format == vibez_project::project_format_v1::ProjectFileFormat::V1,
-                )
-            });
-            let saved =
-                vibez_project::project_format_v1::save_project_v1(&path, v1_source, project)
-                    .map_err(|error| error.to_string())?;
-            Ok(ProjectSaveResult {
-                path,
-                project: saved.project,
-                observation: Some(saved.observation),
-            })
-        } else {
-            // Legacy JSON has no Project Media table; transient staging
-            // references must resolve back to durable Source Storage
-            // identity or they dangle once the staging cache is swept.
-            let mut project = project;
-            vibez_project::project_format_v1::strip_staged_sources(&mut project);
-            project
-                .save_to_file(&path)
-                .map_err(|error| error.to_string())?;
-            Ok(ProjectSaveResult {
-                path,
-                project,
-                observation: None,
-            })
-        }
-    })
-    .await
-    .map_err(|err| format!("save task failed: {err}"))?
-}
-
-pub(super) async fn quantize_audio_clip_async(
-    input: QuantizeInput,
-) -> Result<crate::message::AudioQuantizeSuccess, String> {
-    tokio::task::spawn_blocking(move || compute_audio_quantize(input))
-        .await
-        .map_err(|e| format!("quantize task failed: {e}"))?
-}
-
-pub(super) async fn detect_clip_bpm_async(
-    audio: Arc<vibez_core::audio_buffer::DecodedAudio>,
-    sample_rate: u32,
-) -> Option<vibez_core::onset::BpmEstimate> {
-    tokio::task::spawn_blocking(move || vibez_core::onset::detect_bpm(&audio, sample_rate))
-        .await
-        .unwrap_or(None)
-}
-
-pub(super) async fn detect_clip_transients_async(
-    audio: Arc<vibez_core::audio_buffer::DecodedAudio>,
-    sensitivity: vibez_core::onset::TransientSensitivity,
-) -> Vec<u64> {
-    tokio::task::spawn_blocking(move || vibez_core::onset::detect_onsets(&audio, sensitivity))
-        .await
-        .unwrap_or_default()
-}
-
-pub(super) async fn transpose_clip_async(
-    request: crate::domains::arrangement::ClipTransposeRenderRequest,
-) -> Result<crate::message::ClipTransposeSuccess, String> {
-    run_off_ui_thread("clip transpose", move || {
-        let (audio, transpose_fallback) =
-            vibez_dsp::time_stretch::pitch_preserving_stretch_transposed_checked(
-                &request.source_audio,
-                request.target_frames,
-                f32::from(request.transpose.semitones()),
-            );
-        crate::message::ClipTransposeSuccess {
-            audio: Arc::new(audio),
-            source_audio: request.source_audio,
-            transpose: request.transpose,
-            expected_warped: request.expected_warped,
-            expected_audio: request.expected_audio,
-            expected_geometry: request.expected_geometry,
-            geometry: request.geometry,
-            warning: transpose_fallback.then(|| {
-                "Signalsmith declined this render; timing was restored but Transpose was not applied"
-                    .to_string()
-            }),
-        }
-    })
-    .await
-}
-
-pub(super) async fn warp_browser_audition_async(
-    audio: Arc<vibez_core::audio_buffer::DecodedAudio>,
-    source_bpm: f64,
-    project_bpm: f64,
-) -> Result<Arc<vibez_core::audio_buffer::DecodedAudio>, String> {
-    tokio::task::spawn_blocking(move || {
-        crate::warp::rewarp_for_load(&audio, source_bpm, project_bpm, 0)
-            .ok_or_else(|| "Could not create pitch-preserving WARP Audition".to_string())
-    })
-    .await
-    .map_err(|error| format!("audition warp task failed: {error}"))?
-}
-
-pub(super) async fn prepare_browser_import_audio_async(
-    target: crate::message::BrowserImportTarget,
-    treatment: crate::state::AuditionImportInput,
-    raw: Arc<vibez_core::audio_buffer::DecodedAudio>,
-    source: MediaSourceRef,
-    project_bpm: f64,
-) -> Result<
-    (
-        Arc<vibez_core::audio_buffer::DecodedAudio>,
-        Option<Arc<vibez_core::audio_buffer::DecodedAudio>>,
-        MediaSourceRef,
-    ),
-    String,
-> {
-    if treatment.mode == crate::state::AuditionMode::Raw {
-        return Ok((raw, None, source));
-    }
-    let source_bpm = treatment
-        .source_bpm
-        .filter(|bpm| bpm.is_finite() && *bpm > 0.0)
-        .ok_or_else(|| "Confirm a positive source BPM before WARP import".to_string())?;
-    let frames = raw.num_frames() as u64;
-    let success = crate::warp::warp_clip_async(crate::warp::WarpClipInput {
-        audio: Arc::clone(&raw),
-        fields_frames: frames,
-        source_offset: 0,
-        start_marker: 0,
-        duration: frames,
-        loop_start: 0,
-        loop_end: frames,
-        clip_bpm: source_bpm,
-        project_bpm,
-        transpose_semitones: 0,
-    })
-    .await?;
-    let device_target = matches!(
-        target,
-        crate::message::BrowserImportTarget::Sampler(_)
-            | crate::message::BrowserImportTarget::DrumRackPad { .. }
-    );
-    if !device_target {
-        return Ok((success.audio, Some(success.original_audio), source));
-    }
-
-    let rendered = Arc::clone(&success.audio);
-    let staged = tokio::task::spawn_blocking(move || {
-        let original_name = source.display_name();
-        let stem = std::path::Path::new(&original_name)
-            .file_stem()
-            .map(|stem| stem.to_string_lossy().into_owned())
-            .filter(|stem| !stem.is_empty())
-            .unwrap_or_else(|| "sample".into());
-        let file_name = format!("{stem}-warped.wav");
-        let nonce = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|duration| duration.as_nanos())
-            .unwrap_or_default();
-        let temporary = std::env::temp_dir().join(format!(
-            "vibez-warp-import-{}-{nonce}.wav",
-            std::process::id()
-        ));
-        vibez_audio_io::file_io::write_wav_file(&temporary, &rendered)
-            .map_err(|error| error.to_string())?;
-        let content = std::fs::read(&temporary).map_err(|error| error.to_string())?;
-        let _ = std::fs::remove_file(&temporary);
-        match source {
-            MediaSourceRef::StagedProjectMedia { source_path, .. }
-            | MediaSourceRef::LocalFile { path: source_path } => {
-                vibez_project::project_format_v1::stage_local_content(
-                    &source_path,
-                    &file_name,
-                    &content,
-                )
-                .map_err(|error| error.to_string())
-            }
-            MediaSourceRef::StagedRemoteProjectMedia { provenance, .. } => match *provenance {
-                vibez_core::track::MediaProvenance::Remote {
-                    provider,
-                    connection_id,
-                    connection_name,
-                    source_id,
-                    source_path,
-                    revision,
-                } => vibez_project::project_format_v1::stage_remote_content(
-                    &file_name,
-                    &content,
-                    vibez_core::track::MediaProvenance::Remote {
-                        provider,
-                        connection_id,
-                        connection_name,
-                        source_id,
-                        source_path,
-                        revision,
-                    },
-                )
-                .map_err(|error| error.to_string()),
-                vibez_core::track::MediaProvenance::Local { .. } => {
-                    Err("Remote staging carried Local provenance".to_string())
-                }
-            },
-            _ => Err("WARP device import requires materialized Project Media".to_string()),
-        }
-    })
-    .await
-    .map_err(|error| format!("WARP device staging task failed: {error}"))??;
-    Ok((success.audio, None, staged))
-}
-
-pub(super) async fn auto_warp_clip_async(input: AutoWarpInput) -> crate::message::AutoWarpOutcome {
-    use crate::message::AutoWarpOutcome;
-    let audio_for_detect = Arc::clone(&input.audio);
-    let sample_rate = input.sample_rate;
-    let estimate = tokio::task::spawn_blocking(move || {
-        vibez_core::onset::detect_bpm(&audio_for_detect, sample_rate)
-    })
-    .await
-    .unwrap_or(None);
-    let Some(est) = estimate else {
-        return AutoWarpOutcome::NotDetected;
-    };
-    if est.confidence < input.confidence_threshold || est.bpm <= 0.0 {
-        return AutoWarpOutcome::DetectedOnly {
-            bpm: est.bpm,
-            confidence: est.confidence,
-        };
-    }
-    let num_frames = input.audio.num_frames();
-    let warp_input = crate::warp::WarpClipInput {
-        audio: input.audio,
-        fields_frames: num_frames as u64,
-        source_offset: 0,
-        start_marker: 0,
-        duration: num_frames as u64,
-        loop_start: 0,
-        loop_end: 0,
-        clip_bpm: est.bpm,
-        project_bpm: input.project_bpm,
-        transpose_semitones: 0,
-    };
-    match crate::warp::warp_clip_async(warp_input).await {
-        Ok(success) => AutoWarpOutcome::Warped {
-            confidence: est.confidence,
-            success,
-        },
-        Err(_) => AutoWarpOutcome::DetectedOnly {
-            bpm: est.bpm,
-            confidence: est.confidence,
-        },
-    }
-}
-
-pub(super) async fn connect_dropbox_async(
-    app_key: String,
-) -> Result<(vibez_dropbox::AccountInfo, vibez_dropbox::Tokens), String> {
-    let opener: Arc<dyn vibez_dropbox::BrowserOpener> =
-        Arc::new(vibez_dropbox::SystemBrowserOpener);
-    let tokens = vibez_dropbox::run_oauth_flow(&app_key, opener)
-        .await
-        .map_err(|e| e.to_string())?;
-    let client = DropboxClient::new(app_key, tokens);
-    let info = client.current_account().await.map_err(|e| e.to_string())?;
-    let tokens = client.tokens().await;
-    Ok((info, tokens))
-}
-
-/// Hand a URL to the system browser on a blocking thread.
-///
-/// Reuses the opener vibez-dropbox already owns for its OAuth flow rather
-/// than taking a second browser dependency. `BrowserOpener::open` waits for
-/// the platform launcher to return, which must never happen on the update
-/// loop: on a cold browser start that stall would freeze the whole UI.
-pub(super) async fn open_url_async(url: &'static str) -> Result<(), String> {
-    let opener: Arc<dyn vibez_dropbox::BrowserOpener> =
-        Arc::new(vibez_dropbox::SystemBrowserOpener);
-    tokio::task::spawn_blocking(move || opener.open(url))
-        .await
-        .map_err(|error| format!("Browser task failed: {error}"))?
-        .map_err(|error| error.to_string())
-}
-
-/// Commit downloaded bytes to the Media Cache on a blocking thread; the
-/// write can be multi-MB and must not stall the async executor.
-pub(super) async fn write_cache_blocking(
-    cache: &DropboxCache,
-    entry: &DropboxEntry,
-    bytes: Vec<u8>,
-) -> Result<PathBuf, String> {
-    let cache = cache.clone();
-    let path_lower = entry.path_lower.clone();
-    let revision = entry.rev.clone();
-    tokio::task::spawn_blocking(move || cache.write(&path_lower, revision.as_deref(), &bytes))
-        .await
-        .map_err(|error| format!("Media Cache write task failed: {error}"))?
-        .map_err(|error| format!("Media Cache write failed: {error}"))
-}
-
-pub(super) async fn fetch_dropbox_sample_async(
-    client: Option<Arc<DropboxClient>>,
-    cache: DropboxCache,
-    entry: DropboxEntry,
-) -> Result<(crate::message::AnalysedBrowserAudio, String, MediaSourceRef), String> {
-    let _lease = cache.protect(&entry.path_lower, entry.rev.as_deref());
-    let local = match cache
-        .lookup(&entry.path_lower, entry.rev.as_deref())
-        .map_err(|error| format!("Media Cache lookup failed: {error}"))?
-    {
-        Some(path) => path,
-        None => {
-            let client = client.ok_or_else(|| {
-                "Reconnect Required · uncached Remote media cannot be imported".to_string()
-            })?;
-            let bytes = client.download(&entry.path_lower).await.map_err(|error| {
-                format!("Remote materialization failed for {}: {error}", entry.name)
-            })?;
-            write_cache_blocking(&cache, &entry, bytes).await?
-        }
-    };
-    let metadata_cache = cache.clone();
-    let metadata_path = entry.path_lower.clone();
-    let metadata_revision = entry.rev.clone();
-    let cached_metadata = tokio::task::spawn_blocking(move || {
-        metadata_cache.derived_metadata(&metadata_path, metadata_revision.as_deref())
-    })
-    .await
-    .map_err(|error| format!("Derived Metadata lookup task failed: {error}"))?
-    .map_err(|error| format!("Derived Metadata lookup failed: {error}"))?;
-    let decode_path = local.clone();
-    let decoded = tokio::task::spawn_blocking(move || {
-        vibez_audio_io::file_io::decode_audio_file(&decode_path).map_err(|e| e.to_string())
-    })
-    .await
-    .map_err(|e| format!("decode task failed: {e}"))??;
-    let staging_entry = entry.clone();
-    let source = tokio::task::spawn_blocking(move || {
-        vibez_project::project_format_v1::stage_remote_file(
-            &local,
-            &staging_entry.name,
-            vibez_core::track::MediaProvenance::Remote {
-                provider: crate::remote_provider::DROPBOX_PROVIDER_ID.into(),
-                connection_id: crate::remote_provider::DROPBOX_CONNECTION_ID.into(),
-                connection_name: Some(crate::remote_provider::DROPBOX_CONNECTION_NAME.into()),
-                source_id: staging_entry.path_lower,
-                source_path: staging_entry.path_display,
-                revision: staging_entry.rev,
-            },
-        )
-        .map_err(|error| format!("Remote Project Media staging failed: {error}"))
-    })
-    .await
-    .map_err(|error| format!("Remote Project Media staging task failed: {error}"))??;
-    let analysed = analyse_browser_audio_with_cached_metadata_async(
-        decoded,
-        entry.name.clone(),
-        cached_metadata,
-    )
-    .await?;
-    Ok((analysed, entry.name, source))
-}
-
-pub(super) async fn materialize_remote_sample_async(
-    client: Option<Arc<DropboxClient>>,
-    cache: DropboxCache,
-    entry: DropboxEntry,
-    lease: vibez_dropbox::CacheLease,
-    debounce: bool,
-) -> Result<crate::message::RemoteMaterializedSample, String> {
-    if debounce
-        && cache
-            .lookup(&entry.path_lower, entry.rev.as_deref())
-            .map_err(|error| format!("Media Cache lookup failed: {error}"))?
-            .is_none()
-    {
-        tokio::time::sleep(dropbox_io::REMOTE_SELECTION_DEBOUNCE).await;
-    }
-
-    let local = match cache
-        .lookup(&entry.path_lower, entry.rev.as_deref())
-        .map_err(|error| format!("Media Cache lookup failed: {error}"))?
-    {
-        Some(path) => path,
-        None => {
-            let client = client.ok_or_else(|| {
-                "Reconnect Required · uncached Remote media cannot be materialized".to_string()
-            })?;
-            let bytes = client.download(&entry.path_lower).await.map_err(|error| {
-                format!("Remote materialization failed for {}: {error}", entry.name)
-            })?;
-            write_cache_blocking(&cache, &entry, bytes).await?
-        }
-    };
-
-    let revision = entry.rev.clone();
-    let (decoded, metadata) = tokio::task::spawn_blocking(move || {
-        let decoded = vibez_audio_io::file_io::decode_audio_file(&local)
-            .map_err(|error| error.to_string())?;
-        let estimate = vibez_core::onset::detect_bpm(&decoded, decoded.sample_rate);
-        let bucket_count = 64usize;
-        let frames_per_bucket = decoded.num_frames().max(1).div_ceil(bucket_count);
-        let waveform_peaks = (0..bucket_count)
-            .map(|bucket| {
-                let start = bucket * frames_per_bucket;
-                let end = (start + frames_per_bucket).min(decoded.num_frames());
-                (0..decoded.num_channels())
-                    .map(|channel| {
-                        let (min, max) = decoded.peak_in_range(channel, start, end);
-                        min.abs().max(max.abs())
-                    })
-                    .fold(0.0_f32, f32::max)
-            })
-            .collect();
-        let metadata = vibez_dropbox::DerivedMetadata {
-            provider_revision: revision,
-            duration_seconds: decoded.duration_seconds(),
-            channels: decoded.num_channels().try_into().unwrap_or(u16::MAX),
-            sample_rate: decoded.sample_rate,
-            bpm: estimate.map(|value| value.bpm),
-            bpm_confidence: estimate.map(|value| value.confidence),
-            waveform_peaks,
-        };
-        Ok::<_, String>((Arc::new(decoded), metadata))
-    })
-    .await
-    .map_err(|error| format!("decode task failed: {error}"))??;
-    cache
-        .store_derived_metadata(&entry.path_lower, entry.rev.as_deref(), metadata.clone())
-        .map_err(|error| format!("Derived Metadata save failed: {error}"))?;
-    let source = MediaSourceRef::DropboxFile {
-        path_lower: entry.path_lower,
-        display_path: entry.path_display,
-        rev: entry.rev,
-    };
-    Ok(crate::message::RemoteMaterializedSample {
-        audio: decoded,
-        name: entry.name,
-        source,
-        lease,
-        metadata,
-    })
-}
-
-pub(super) async fn export_async(
-    request: vibez_engine::render::BounceRequest,
-    mut plugins: vibez_engine::render::OfflinePlugins,
-    wav_path: PathBuf,
-    progress: Arc<std::sync::atomic::AtomicU8>,
-    plugin_return: std::sync::mpsc::Sender<vibez_engine::render::OfflinePlugins>,
-) -> Result<PathBuf, String> {
-    tokio::task::spawn_blocking(move || {
-        use std::sync::atomic::Ordering;
-
-        let outcome = (|| {
-            let result = vibez_engine::render::render_offline_with_plugins(
-                &request,
-                &mut plugins,
-                |percent| {
-                    // Plugin preflight occupies 0–10%; audio rendering occupies
-                    // 10–99%, with 100 reserved for the committed destination.
-                    let scaled = 10 + ((percent as u16 * 89) / 100) as u8;
-                    progress.store(scaled.min(99), Ordering::Relaxed);
-                },
-            )?;
-            if !result.warnings.is_empty() {
-                return Err(format!(
-                    "render was incomplete: {}",
-                    result.warnings.join("; ")
-                ));
-            }
-            let temporary = temporary_export_path(&wav_path);
-            if let Err(error) = vibez_audio_io::file_io::write_wav_file(&temporary, &result.audio) {
-                let _ = std::fs::remove_file(&temporary);
-                return Err(format!("WAV write error: {error}"));
-            }
-            if let Err(error) = std::fs::rename(&temporary, &wav_path) {
-                let _ = std::fs::remove_file(&temporary);
-                return Err(format!("could not commit destination WAV: {error}"));
-            }
-            progress.store(100, Ordering::Relaxed);
-            Ok(wav_path)
-        })();
-        // Plugin teardown is main-thread-affine for JUCE-based CLAP/VST3
-        // devices. Return the instances to App; dropping them here would
-        // violate the plugin contract after an otherwise successful export.
-        let _ = plugin_return.send(plugins);
-        outcome
-    })
-    .await
-    .map_err(|err| format!("export task failed: {err}"))?
-}
-
-fn temporary_export_path(destination: &std::path::Path) -> PathBuf {
-    let nonce = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|duration| duration.as_nanos())
-        .unwrap_or_default();
-    let file_name = destination
-        .file_name()
-        .map(|name| name.to_string_lossy().into_owned())
-        .unwrap_or_else(|| "vibez-export.wav".to_string());
-    destination.with_file_name(format!(".{file_name}.{nonce}.part"))
-}
-
-pub(super) async fn bounce_async(
-    request: vibez_engine::render::BounceRequest,
-    wav_path: PathBuf,
-    clip_name: String,
-    insert_position_samples: u64,
-) -> Result<crate::message::BounceOutcome, String> {
-    tokio::task::spawn_blocking(move || {
-        let result = vibez_engine::render::render_offline(&request);
-        vibez_audio_io::file_io::write_wav_file(&wav_path, &result.audio)
-            .map_err(|e| format!("WAV write error: {e}"))?;
-        Ok(crate::message::BounceOutcome {
-            audio: Arc::new(result.audio),
-            source: MediaSourceRef::LocalFile {
-                path: wav_path.clone(),
-            },
-            path: wav_path,
-            clip_name,
-            insert_position_samples,
-            warnings: result.warnings,
-        })
-    })
-    .await
-    .map_err(|err| format!("bounce task failed: {err}"))?
-}
-
-/// Finish a decoded clip for project load. The project file stores
-/// the raw source reference, but a warped clip's geometry (duration /
-/// offsets / loop bounds) is saved in warped-sample units, so the
-/// deterministic stretch is re-applied here; otherwise every warped
-/// clip reloads at its raw tempo and the whole project plays out of
-/// sync. The stretch runs on a blocking thread (WSOLA over a whole
-/// clip is CPU-heavy).
-pub(super) async fn finish_loaded_clip(
-    info: ClipInfo,
-    raw: Arc<vibez_core::audio_buffer::DecodedAudio>,
-) -> LoadedClipData {
-    let transpose = info.transpose.semitones();
-    if info.warped {
-        if let (Some(clip_bpm), Some(warped_to_bpm)) = (info.original_bpm, info.warped_to_bpm) {
-            let stretch_src = Arc::clone(&raw);
-            let warped = tokio::task::spawn_blocking(move || {
-                crate::warp::rewarp_for_load(&stretch_src, clip_bpm, warped_to_bpm, transpose)
-            })
-            .await
-            .unwrap_or(None);
-            if let Some(warped) = warped {
-                return LoadedClipData {
-                    info,
-                    audio: warped,
-                    original_audio: Some(raw),
-                };
-            }
-        }
-    }
-    if transpose != 0 {
-        let source = Arc::clone(&raw);
-        let rendered = tokio::task::spawn_blocking(move || {
-            Arc::new(
-                vibez_dsp::time_stretch::pitch_preserving_stretch_transposed(
-                    &source,
-                    source.num_frames(),
-                    f32::from(transpose),
-                ),
-            )
-        })
-        .await
-        .unwrap_or_else(|_| Arc::clone(&raw));
-        return LoadedClipData {
-            info,
-            audio: rendered,
-            original_audio: Some(raw),
-        };
-    }
-    LoadedClipData {
-        info,
-        audio: raw,
-        original_audio: None,
-    }
-}
-
-/// Run CPU-bound work on the tokio blocking pool, labelling join failures.
-/// Every off-UI-thread hop in the app routes through here so the idiom (and
-/// its error shape) exists exactly once.
-pub(crate) async fn run_off_ui_thread<T, F>(label: &'static str, work: F) -> Result<T, String>
-where
-    T: Send + 'static,
-    F: FnOnce() -> T + Send + 'static,
-{
-    tokio::task::spawn_blocking(work)
-        .await
-        .map_err(|error| format!("{label} task failed: {error}"))
-}
-
-/// One rule for what counts as a missing project: only a confirmed
-/// `NotFound` io error prunes a Recent Projects entry; everything else is
-/// treated as transient.
-fn classify_load_error(
-    io_error: Option<&std::io::Error>,
-    message: String,
-) -> crate::message::ProjectLoadError {
-    if io_error.is_some_and(|error| error.kind() == std::io::ErrorKind::NotFound) {
-        crate::message::ProjectLoadError::missing_project(message)
-    } else {
-        crate::message::ProjectLoadError::other(message)
-    }
-}
-
-fn classify_project_format_load_error(
-    error: vibez_project::project_format_v1::ProjectFormatError,
-) -> crate::message::ProjectLoadError {
-    let io_error = match &error {
-        vibez_project::project_format_v1::ProjectFormatError::Io(io_error) => Some(io_error),
-        _ => None,
-    };
-    classify_load_error(io_error, error.to_string())
-}
-
-fn classify_legacy_project_load_error(
-    error: vibez_project::ProjectError,
-) -> crate::message::ProjectLoadError {
-    let io_error = match &error {
-        vibez_project::ProjectError::Io(io_error) => Some(io_error),
-        _ => None,
-    };
-    classify_load_error(io_error, error.to_string())
-}
-
-pub(super) async fn load_project_async(
-    path: PathBuf,
-    dropbox: Option<(Arc<DropboxClient>, DropboxCache)>,
-) -> Result<ProjectLoadResult, crate::message::ProjectLoadError> {
-    let load_path = path.clone();
-    let (project, container_path) = tokio::task::spawn_blocking(move || {
-        match vibez_project::project_format_v1::detect_project_format(&load_path)
-            .map_err(classify_project_format_load_error)?
-        {
-            vibez_project::project_format_v1::ProjectFileFormat::V1 => {
-                let container =
-                    vibez_project::project_format_v1::ProjectContainer::open(&load_path)
-                        .map_err(classify_project_format_load_error)?;
-                let document = container
-                    .load_document()
-                    .map_err(classify_project_format_load_error)?;
-                Ok((document.project, Some(load_path)))
-            }
-            vibez_project::project_format_v1::ProjectFileFormat::LegacyJson => {
-                Project::load_from_file(&load_path)
-                    .map(|project| (project, None))
-                    .map_err(classify_legacy_project_load_error)
-            }
-        }
-    })
-    .await
-    .map_err(|err| crate::message::ProjectLoadError::other(format!("load task failed: {err}")))??;
-
-    let mut clips = Vec::new();
-    let mut unresolved_clips = Vec::new();
-    let mut sampler_samples = Vec::new();
-    let mut drum_rack_pad_samples = Vec::new();
-    let mut warnings = Vec::new();
-
-    for (location, timeline) in project.timelines() {
-        for clip in &timeline.clips {
-            match clip.resolved_source().cloned() {
-                Some(source) => match hydrate_saved_source(
-                    container_path.as_ref(),
-                    dropbox.as_ref(),
-                    &source,
-                    &clip.name,
-                )
-                .await
-                {
-                    Ok(audio) => clips.push(crate::message::LoadedTimelineClip {
-                        location,
-                        clip: finish_loaded_clip(clip.clone(), Arc::new(audio)).await,
-                    }),
-                    Err(err) => {
-                        warnings.push(format!(
-                            "Clip '{}' unavailable, kept for relink ({})",
-                            clip.name, err
-                        ));
-                        unresolved_clips.push(crate::message::UnresolvedTimelineClip {
-                            location,
-                            info: clip.clone(),
-                        });
-                    }
-                },
-                None => warnings.push(format!(
-                    "Skipped clip '{}' (missing source reference)",
-                    clip.name
-                )),
-            }
-        }
-    }
-
-    for track in &project.tracks {
-        if let Some(native) = &track.native_instrument {
-            match native {
-                InstrumentStateInfo::Sampler {
-                    source: Some(source),
-                    ..
-                } => match hydrate_saved_source(
-                    container_path.as_ref(),
-                    dropbox.as_ref(),
-                    source,
-                    &track.name,
-                )
-                .await
-                {
-                    Ok(audio) => sampler_samples.push(LoadedSamplerData {
-                        track_id: track.id,
-                        source: source.clone(),
-                        audio: Arc::new(audio),
-                        name: source.display_name(),
-                    }),
-                    Err(err) => warnings.push(format!(
-                        "Skipped sampler source on '{}' ({})",
-                        track.name, err
-                    )),
-                },
-                InstrumentStateInfo::DrumRack { pads } => {
-                    for (pad_index, pad) in pads.iter().enumerate() {
-                        let Some(source) = &pad.source else {
-                            continue;
-                        };
-                        let label = format!("drum pad {} on '{}'", pad_index + 1, track.name);
-                        match hydrate_saved_source(
-                            container_path.as_ref(),
-                            dropbox.as_ref(),
-                            source,
-                            &label,
-                        )
-                        .await
-                        {
-                            Ok(audio) => drum_rack_pad_samples.push(LoadedDrumRackPadData {
-                                track_id: track.id,
-                                pad_index,
-                                source: source.clone(),
-                                audio: Arc::new(audio),
-                                name: pad.name.clone().unwrap_or_else(|| source.display_name()),
-                                state: pad.clone(),
-                            }),
-                            Err(err) => warnings.push(format!("Skipped {label} ({err})")),
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-
-    Ok(ProjectLoadResult {
-        path,
-        project,
-        clips,
-        unresolved_clips,
-        sampler_samples,
-        drum_rack_pad_samples,
-        warnings,
-    })
-}
-
-pub(super) async fn hydrate_saved_source(
-    container_path: Option<&PathBuf>,
-    dropbox: Option<&(Arc<DropboxClient>, DropboxCache)>,
-    source: &MediaSourceRef,
-    label: &str,
-) -> Result<vibez_core::audio_buffer::DecodedAudio, String> {
-    match source {
-        MediaSourceRef::LocalFile { path }
-        | MediaSourceRef::StagedProjectMedia {
-            staging_path: path, ..
-        }
-        | MediaSourceRef::StagedRemoteProjectMedia {
-            staging_path: path, ..
-        } => decode_blocking(path.clone()).await,
-        MediaSourceRef::ProjectMedia { id, file_name, .. } => {
-            let container_path = container_path
-                .cloned()
-                .ok_or_else(|| format!("{label} has Project Media without a V1 container"))?;
-            let id = id.clone();
-            let extension = Path::new(file_name)
-                .extension()
-                .map(|value| value.to_string_lossy().into_owned());
-            tokio::task::spawn_blocking(move || {
-                let container =
-                    vibez_project::project_format_v1::ProjectContainer::open(container_path)
-                        .map_err(|error| error.to_string())?;
-                let bytes = container
-                    .read_media(&id)
-                    .map_err(|error| error.to_string())?;
-                vibez_audio_io::file_io::decode_audio_cursor(
-                    std::io::Cursor::new(bytes),
-                    extension.as_deref(),
-                )
-                .map_err(|error| error.to_string())
-            })
-            .await
-            .map_err(|error| format!("Project Media decode task failed: {error}"))?
-        }
-        MediaSourceRef::DropboxFile { .. } => hydrate_dropbox_source(dropbox, source, label).await,
-    }
-}
-
-pub(super) async fn decode_blocking(
-    path: PathBuf,
-) -> Result<vibez_core::audio_buffer::DecodedAudio, String> {
-    tokio::task::spawn_blocking(move || {
-        file_io::decode_audio_file(&path).map_err(|e| e.to_string())
-    })
-    .await
-    .map_err(|e| format!("decode task failed: {e}"))?
-}
-
-pub(super) async fn hydrate_dropbox_source(
-    dropbox: Option<&(Arc<DropboxClient>, DropboxCache)>,
-    source: &MediaSourceRef,
-    label: &str,
-) -> Result<vibez_core::audio_buffer::DecodedAudio, String> {
-    let MediaSourceRef::DropboxFile {
-        path_lower,
-        display_path,
-        rev,
-    } = source
-    else {
-        return Err(format!(
-            "Skipped '{label}' (expected Dropbox source reference)"
-        ));
-    };
-    let Some((client, cache)) = dropbox else {
-        return Err(format!(
-            "Skipped '{label}' (not connected to Dropbox - reconnect in Settings)"
-        ));
-    };
-    let file_name = display_path
-        .rsplit_once('/')
-        .map(|(_, n)| n.to_string())
-        .unwrap_or_else(|| display_path.clone());
-    let entry = DropboxEntry {
-        path_lower: path_lower.clone(),
-        path_display: display_path.clone(),
-        name: file_name,
-        is_folder: false,
-        rev: rev.clone(),
-        size: None,
-    };
-    let local_path = client
-        .download_to_cache(&entry, cache)
-        .await
-        .map_err(|e| format!("Skipped '{label}' ({e})"))?;
-    decode_blocking(local_path)
-        .await
-        .map_err(|e| format!("Skipped '{label}' ({e})"))
-}
-
-pub(super) async fn scan_sample_root_async(
-    root: PathBuf,
-) -> Result<SampleLibraryScanResult, String> {
-    tokio::task::spawn_blocking(move || scan_sample_root(&root))
-        .await
-        .map_err(|err| format!("scan task failed: {err}"))?
-}
-
-#[cfg(test)]
-mod export_tests {
-    use super::*;
-    use vibez_core::constants::{DEFAULT_TRACK_GAIN, DEFAULT_TRACK_PAN};
-    use vibez_core::effect::PluginDeviceInfo;
-    use vibez_core::id::TrackId;
-    use vibez_core::midi::TrackKind;
-    use vibez_core::track::TrackInfo;
-
-    #[tokio::test]
-    async fn failed_plugin_render_never_creates_the_destination_wav() {
-        let track_id = TrackId::new();
-        let track = TrackInfo {
-            id: track_id,
-            name: "Bass".into(),
-            gain: DEFAULT_TRACK_GAIN,
-            pan: DEFAULT_TRACK_PAN,
-            mute: false,
-            solo: false,
-            audio_input_route: Default::default(),
-            input_monitoring: Default::default(),
-            swing_offset: None,
-            effects: Vec::new(),
-            kind: TrackKind::Midi,
-            color_index: 0,
-            instrument: None,
-            native_instrument: None,
-            plugin_instrument: Some(PluginDeviceInfo {
-                format: "clap".into(),
-                uid: "org.example.bass".into(),
-                path: "/missing/Bass.clap".into(),
-                name: "Bass Plugin".into(),
-                state_b64: None,
-            }),
-            automation: Vec::new(),
-            sends: Vec::new(),
-        };
-        let request = vibez_engine::render::BounceRequest {
-            tracks: vec![track],
-            master: None,
-            buses: Vec::new(),
-            audio_clips: Vec::new(),
-            note_clips: Vec::new(),
-            clip_audio: Default::default(),
-            sampler_audio: Default::default(),
-            drum_pad_audio: Default::default(),
-            mode: vibez_engine::render::BounceMode::Master,
-            range_samples: (0, 512),
-            bpm: 120.0,
-            sample_rate: 44_100,
-            swing: vibez_core::perform::SwingAmount::STRAIGHT,
-        };
-        let directory = tempfile::tempdir().unwrap();
-        let destination = directory.path().join("must-not-exist.wav");
-        let progress = Arc::new(std::sync::atomic::AtomicU8::new(0));
-        let (plugin_return, returned_plugins) = std::sync::mpsc::channel();
-
-        let error = export_async(
-            request,
-            vibez_engine::render::OfflinePlugins::default(),
-            destination.clone(),
-            progress,
-            plugin_return,
-        )
-        .await
-        .unwrap_err();
-        drop(returned_plugins.recv().unwrap());
-
-        assert!(error.contains("Bass Plugin"));
-        assert!(!destination.exists());
-        assert_eq!(
-            std::fs::read_dir(directory.path()).unwrap().count(),
-            0,
-            "failed export must clean up every temporary file"
-        );
-    }
-}
-
-#[cfg(test)]
-mod off_ui_thread_tests {
-    use std::time::Duration;
-
-    use super::run_off_ui_thread;
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn blocking_work_never_blocks_the_ui_executor() {
-        let (release_tx, release_rx) = std::sync::mpsc::channel();
-        let work = tokio::spawn(run_off_ui_thread("Test", move || {
-            release_rx
-                .recv_timeout(Duration::from_secs(1))
-                .expect("UI executor should release the blocking worker");
-            42
-        }));
-
-        tokio::task::yield_now().await;
-        assert!(release_tx.send(()).is_ok());
-        assert_eq!(work.await.unwrap().unwrap(), 42);
-    }
-}
-
-#[cfg(test)]
-mod project_load_error_tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn a_missing_top_level_project_is_classified_for_recent_path_pruning() {
-        let directory = tempfile::tempdir().unwrap();
-        let missing = directory.path().join("removed.vzp");
-
-        let error = load_project_async(missing, None).await.unwrap_err();
-
-        assert!(matches!(
-            error,
-            crate::message::ProjectLoadError::MissingProject(_)
-        ));
-    }
-
-    #[tokio::test]
-    async fn an_existing_but_invalid_project_is_not_classified_as_missing() {
-        let directory = tempfile::tempdir().unwrap();
-        let invalid = directory.path().join("invalid.vzp");
-        std::fs::write(&invalid, "not a Vibez project").unwrap();
-
-        let error = load_project_async(invalid, None).await.unwrap_err();
-
-        assert!(matches!(error, crate::message::ProjectLoadError::Other(_)));
-    }
-}
