+++ crates/vibez-ui/src/app/mod.rs
@@ -153,7 +153,6 @@ pub fn run() -> iced::Result {
 }
 
 mod actions;
-mod async_helpers;
 mod audio_recording;
 mod audio_settings;
 mod audio_take_finalization;
@@ -165,12 +164,13 @@ mod engine_events;
 mod keyboard;
 mod local_watcher;
 mod menu_lifecycle;
+mod tasks;
 mod tracked_request;
 
-use async_helpers::*;
 pub(crate) use audio_tasks::*;
 use drum_rack_preparation::*;
 pub(crate) use keyboard::*;
+use tasks::*;
 mod dropbox_io;
 mod media;
 mod media_import;
